add Resize functions
We want users to be able to specify when and how much to reduce pool
sizes by, and to do so without locking and unlocking the data structure
for each item removed.

So we provide a Resize function which takes a function that takes the
current size of the pool and returns a new target size. Resize calls
this function, resizes the pool and returns the original size of the
pool, the new size after possibly removing items, and an error if
something went wrong.

This way, users can decide on the size of the reduction and when to
schedule the next Resize call.